extract case update code to separate class
