+++ corehq/ex-submodules/casexml/apps/case/tests/test_rebuild.py
@@ -5,10 +5,11 @@ from casexml.apps.case import const
 from casexml.apps.case.cleanup import rebuild_case_from_forms
 from casexml.apps.case.exceptions import MissingServerDate
 from casexml.apps.case.mock import CaseBlock
-from casexml.apps.case.models import CommCareCase, CommCareCaseAction, _action_sort_key_function
+from casexml.apps.case.models import CommCareCase, CommCareCaseAction
 from datetime import datetime, timedelta
 from copy import deepcopy
 from casexml.apps.case.tests.util import post_util as real_post_util, delete_all_cases
+from casexml.apps.case.update_strategy import _action_sort_key_function, ActionsUpdateStrategy
 from casexml.apps.case.util import primary_actions
 from corehq.form_processor.interfaces import FormProcessorInterface
 from couchforms.models import XFormInstance
@@ -93,7 +94,7 @@ class CaseRebuildTest(TestCase):
         # rebuild by flipping the actions
         case.actions = [case.actions[0], a2, a1]
         case.xform_ids = [case.xform_ids[0], case.xform_ids[2], case.xform_ids[1]]
-        case.rebuild()
+        ActionsUpdateStrategy(case).soft_rebuild_case()
         self.assertEqual(case.p1, 'p1-1') # original
         self.assertEqual(case.p2, 'p2-1') # updated (back!)
         self.assertEqual(case.p3, 'p3-2') # new
@@ -121,19 +122,20 @@ class CaseRebuildTest(TestCase):
         _confirm_action_order(case, [a1, a2, a3])
 
         # test initial rebuild does nothing
-        case.rebuild()
+        update_strategy = ActionsUpdateStrategy(case)
+        update_strategy.soft_rebuild_case()
         _confirm_action_order(case, [a1, a2, a3])
 
         # test sorting by server date
         case.actions[2].server_date = case.actions[2].server_date + timedelta(days=1)
-        case.rebuild()
+        update_strategy.soft_rebuild_case()
         _confirm_action_order(case, [a1, a3, a2])
 
         # test sorting by date within the same day
         case = CommCareCase.get(case_id)
         _confirm_action_order(case, [a1, a2, a3])
         case.actions[2].date = case.actions[3].date + timedelta(minutes=1)
-        case.rebuild()
+        ActionsUpdateStrategy(case).soft_rebuild_case()
         _confirm_action_order(case, [a1, a3, a2])
 
         # test original form order
@@ -141,13 +143,13 @@ class CaseRebuildTest(TestCase):
         case.actions[3].server_date = case.actions[2].server_date
         case.actions[3].date = case.actions[2].date
         case.xform_ids = [a1.xform_id, a3.xform_id, a2.xform_id]
-        case.rebuild()
+        ActionsUpdateStrategy(case).soft_rebuild_case()
         _confirm_action_order(case, [a1, a3, a2])
 
         # test create comes before update
         case = CommCareCase.get(case_id)
         case.actions = [a1, create, a2, a3]
-        case.rebuild()
+        ActionsUpdateStrategy(case).soft_rebuild_case()
         _confirm_action_order(case, [a1, a2, a3])
 
     def testRebuildEmpty(self):
@@ -189,6 +191,7 @@ class CaseRebuildTest(TestCase):
         post_util(case_id=case_id, p1='p1-1', p2='p2-1', form_extras={'received_on': now + timedelta(seconds=1)})
         post_util(case_id=case_id, p2='p2-2', p3='p3-2', form_extras={'received_on': now + timedelta(seconds=2)})
         case = CommCareCase.get(case_id)
+        update_strategy = ActionsUpdateStrategy(case)
 
         original_actions = [deepcopy(a) for a in case.actions]
         original_form_ids = [id for id in case.xform_ids]
@@ -199,21 +202,21 @@ class CaseRebuildTest(TestCase):
         # test reordering
         case.actions = [case.actions[2], case.actions[1], case.actions[0]]
         self._assertListNotEqual(original_actions, case.actions)
-        case.reconcile_actions()
+        update_strategy.reconcile_actions()
         self._assertListEqual(original_actions, case.actions)
 
         # test duplication
         case.actions = case.actions * 3
         self.assertEqual(9, len(case.actions))
         self._assertListNotEqual(original_actions, case.actions)
-        case.reconcile_actions()
+        update_strategy.reconcile_actions()
         self._assertListEqual(original_actions, case.actions)
 
         # test duplication, even when dates are off
         case.actions = original_actions + [deepcopy(case.actions[2])]
         case.actions[-1].server_date = case.actions[-1].server_date + timedelta(seconds=1)
         self._assertListNotEqual(original_actions, case.actions)
-        case.reconcile_actions()
+        update_strategy.reconcile_actions()
         self._assertListEqual(original_actions, case.actions)
 
         # test duplication with different properties is actually
@@ -222,7 +225,7 @@ class CaseRebuildTest(TestCase):
         case.actions[-1].updated_unknown_properties['new'] = 'mismatch'
         self.assertEqual(4, len(case.actions))
         self._assertListNotEqual(original_actions, case.actions)
-        case.reconcile_actions()
+        update_strategy.reconcile_actions()
         self._assertListNotEqual(original_actions, case.actions)
 
         # test clean slate rebuild
@@ -403,7 +406,7 @@ class TestCheckActionOrder(SimpleTestCase):
             CommCareCaseAction(server_date=datetime(2001, 1, 2, 0, 0, 0)),
             CommCareCaseAction(server_date=datetime(2001, 1, 3, 0, 0, 0)),
         ])
-        self.assertTrue(case.check_action_order())
+        self.assertTrue(ActionsUpdateStrategy(case).check_action_order())
 
     def test_out_of_order(self):
         case = CommCareCase(actions=[
@@ -411,7 +414,7 @@ class TestCheckActionOrder(SimpleTestCase):
             CommCareCaseAction(server_date=datetime(2001, 1, 3, 0, 0, 0)),
             CommCareCaseAction(server_date=datetime(2001, 1, 2, 0, 0, 0)),
         ])
-        self.assertFalse(case.check_action_order())
+        self.assertFalse(ActionsUpdateStrategy(case).check_action_order())
 
     def test_sorted_with_none(self):
         case = CommCareCase(actions=[
@@ -420,7 +423,7 @@ class TestCheckActionOrder(SimpleTestCase):
             CommCareCaseAction(server_date=datetime(2001, 1, 2, 0, 0, 0)),
             CommCareCaseAction(server_date=datetime(2001, 1, 3, 0, 0, 0)),
         ])
-        self.assertTrue(case.check_action_order())
+        self.assertTrue(ActionsUpdateStrategy(case).check_action_order())
 
     def test_out_of_order_with_none(self):
         case = CommCareCase(actions=[
@@ -429,7 +432,7 @@ class TestCheckActionOrder(SimpleTestCase):
             CommCareCaseAction(server_date=None),
             CommCareCaseAction(server_date=datetime(2001, 1, 2, 0, 0, 0)),
         ])
-        self.assertFalse(case.check_action_order())
+        self.assertFalse(ActionsUpdateStrategy(case).check_action_order())
 
 
 class TestActionSortKey(SimpleTestCase):
