+++ corehq/ex-submodules/casexml/apps/case/xform.py
@@ -9,6 +9,7 @@ from django.db.models import Q
 import redis
 from casexml.apps.case.dbaccessors import get_reverse_indexed_cases
 from casexml.apps.case.signals import cases_received, case_post_save
+from casexml.apps.case.update_strategy import ActionsUpdateStrategy
 from casexml.apps.phone.cleanliness import should_track_cleanliness, should_create_flags_on_submission
 from casexml.apps.phone.models import OwnershipCleanlinessFlag
 from corehq.toggles import LOOSE_SYNC_TOKEN_VALIDATION
@@ -173,11 +174,7 @@ def process_cases_with_casedb(xforms, case_db, config=None):
         )
 
     for case in cases:
-        if not case.check_action_order():
-            try:
-                case.reconcile_actions(rebuild=True, xforms={xform._id: xform})
-            except ReconciliationError:
-                pass
+        ActionsUpdateStrategy(case).reconcile_actions_if_necessary(xform)
         case_db.mark_changed(case)
 
         action_xforms = {action.xform_id for action in case.actions if action.xform_id}
@@ -402,11 +399,11 @@ def _get_or_update_model(case_update, xform, case_db):
     """
     case = case_db.get(case_update.id)
     if case is None:
-        case = CommCareCase.from_case_update(case_update, xform)
+        case = ActionsUpdateStrategy.case_from_case_update(case_update, xform)
         case_db.set(case['_id'], case)
         return case
     else:
-        case.update_from_case_update(case_update, xform, case_db.get_cached_forms())
+        ActionsUpdateStrategy(case).update_from_case_update(case_update, xform, case_db.get_cached_forms())
         return case
